Capture placeholders and missing object, don't warn (#509)
Right now, placeholders and missing objects is kinda expected behaviour.
If it _were_ a big issue, we wouldn't be consistently shrugging off the
warnings and errors in the console. Having lots of non-actionable
warnings and errors in the console is bad, cause we learn to just ignore
them. We should save these for real problems that require immediate
action.

So I'm making switching these logs to debug level so they're easy to
filter out, but still capturing messages in sentry so we can keep tabs
on how often this is happening.