Hot fix for graph loading (#514)
Pretty sure we're getting crashes in prod from the
`isNewRelatedObjectsPublic` code trying to update a non-node here

https://ideaflow.sentry.io/issues/5964105966/?project=4507929695617024&query=error.unhandled%3Atrue%20is%3Aunresolved&referrer=issue-stream&sort=date&statsPeriod=14d&stream_index=0

Immediate easy fix is to just insist that the `to` is an instance of
`GraphNode`. This isn't quite what we want as overall product behavior,
but we need to come back to this feature for full `GraphObject`
polymorphism later anyway so I'm ok with this for right now.