new pin unpin toggle for node ux (#517)
I've addressed Jacob's desire to streamline the pin-unpin workflow into
a single click. So, in this PR I enhanced the pin button, turning it
into a toggle that appears on hover.

- I changed the UX that always shoed the pin icon to just show on hover.
- implemented more CSS tooltips
- there's some refactor on styling code around the node I had quickly
implemented in the past. This includes improving the tooltip's
reactivity and adding a new tooltip for the pinning toggle.

Completes ENT-4125, ENT-4119

How to test
Just create a few nodes and try out the pin-unpin toggle on the right
side of each node. You'll see it in both the pinned and default
sections.