Improve selection granularity (#416)
## Summary

In our tree selection object, we now store the anchor and focus
positions. In lexical, these are stored as offsets from lexical nodes,
specified by node keys. Node keys are internal implementation details of
lexical, and I'd like to make our app agnostic to those if we can. So we
store our selection anchor/focus as integer offset from the start of an
editor. This also makes our selection resilient to large changes in the
content of the node. As long as the integer positions of our app
selection fall within the bounds of the new content, we can re-apply it.

Roughly how the syncing of selection goes:
- Whenever the editor's selection changes, we update the tree selection
to match
- Whenever the tree selection changes, we update the editor selection to
match. But only if there's a difference between the two. This is what
avoids circular updates. (Note: this feels a little brittle to me but
I'm hopeful is good enough for now. See below for an alternative
approach)

## Changes

- Delete BindFocusToTreePlugin and move it's responsibilities into the
sync plugin. Syncing the selection and content often go hand in hand and
I found it easier to reason about when they're together.
- Renamed SyncWithGraphPlugin to SyncWithModelsPlugin cause it's syncing
with both the graph and tree now. (This means you can't compare the diff
as easily but enough changed that reviewing afresh is warranted
anyways). The plugin now has 3 responsibilities:
- When the editor is updated by user input, we apply the changes to the
graph and tree selection
- When the content of the respective graph node changes, we update the
editor to change and restore the selection by applying the tree
selection
  - When the tree selection changes, we apply it to the editor.
- Added a few helper functions to get/set lexical selection and map b/w
lexical and tree selection
- Misc other renaming for clarity
- In a couple places I made it so setting focus/selection was only done
after a graph update promise resolves (not sure why this wasn't an issue
before tbh)

Completes ENT-3847

## How to test

- Paste a link into a node. Place the caret in the middle of the link
and add/delete something. The caret should stay where you'd expect.
(Previously it would jump to the end).
- Check there's no regressions on existing features which move the
selection:
  - tab
  - up/down
  - split
  - maintain sel after update
  - editor selection
  - add/delete into relation type

## Alternative approach to preventing circular updates

I mentioned above that the approach for preventing circular updates
feels a bit brittle. I had another approach in mind that I didn't
explore, but figure I'd share in case you or I decide we need to try
something else.

- Add a property to the selection object which specifies where it came
from (e.g. a "source" string)
- In general, the plugin doesn't apply selection object states which
come from itself
- Sometimes there's a re-render though in which case we want to restore
the selection, wherever it came from. In this case, we can have an
effect that only runs on mount, which will apply any selection
regardless of source.