Ignore change from global sync channel that's already been received and handled (#541)
Replication steps on main:
- Open app side-by-side, persistence enabled, same account
- Create a new bullet in tabA
- Notice a console warning in tabB

See https://www.loom.com/share/22d43b90149840cf9f8dd5a3e1b2dc7c

When you make a change on tabA, it gets broadcasted out on both the user
channel and global channel. So tabB receives the change twice. The first
one succeeds and the second is rejected.

In this PR, we ignore changes from the same user on the global channel.
These changes will already be received and handled on the user channel.
Note that this is different from the existing filter we have, where
changes from the same *client* are ignored. The fix feels a bit like a
patch, and I suspect there's some better one if I was open to broader
design changes, but I'm not in the mood for that right now.

Part of ENT-4170