Fix indexing of nodes with @ mentions (#538)
Addresses
https://ideaflowteam.slack.com/archives/C07FU15QKTP/p1728353444214029

Replication steps:
- Create node `Global Idea bank`
- Create another `add to @Global Idea bank`
- Create another and trigger autocomplete by pressing ";". Type `add to
global` We want the `add to @Global Idea bank` node to show up but it
doesn't.

We wrap mentions in `@[...]` in the plaintext rendering as a visual
queue to the user that the thing is a mention. This same text is used in
our [keyword
index](https://github.com/IdeaFlowCo/mew/blob/9303b6f2096efef3fe4b0daff47dcbbdc0ed21f0/src/lib/trie.ts#L14).
So in the node `add to @Global Idea bank`, the text "global" doesn't
prefix match against any keywords, so it's filtered out of the results.

The fix is to have not include the `@[...]` in the indexed text.