Fix relation type indexing (#568)
In the relation type dropdown, "child" wasn't showing up. That's because
we were adding the relation type to the index twice, and the second one
with the reverse label was removing the first. Now, we just concatenate
the forward and reverse labels together, which means the relation type
will be indexed across all words in both directions, which is what we
want.

Completes
https://ideaflowteam.slack.com/archives/C07FU15QKTP/p1729489046213889