Add a toast component when creating a new item via the command bar (#573)
do `yarn` to install Toast component

I added toast context, hook, and component. This is used in the command
bar when creating a new node. A link to zoom in on the created node has
been added. We persist the toast on hover.

Completes ENT-4249

How to test: Create a new node through the command bar and watch for the
toast notification to appear in the bottom right.