Taylor/multiline content (#527)
Add multiline notes.

This is a pretty big PR, but it's not quite as scary as the line/file
count suggests. Most of that is migrations, removing dead zone/bundle
code, and renaming.

As noted in the todos below, I haven't done any serious testing of sync.
It includes all the migration and type changes required I think, but I
haven't rigorously tested it.

Feature definition:

- A full walkthrough/explanation of the multiline notes:
https://www.loom.com/share/32a68b24c87b42d3833432a03338ec06
- The original spec has some pictures that are helpful for understanding
the feature, even though the implementation has deviated somewhat:
https://linear.app/ideaflow/issue/ENT-3673/spec-bundles

Code changes:

- Add a new relation list: noteContent. Just like the pinned section is
a subset of all related content, there's now a "note content" section
which related items can be added to. This section is kinda special cause
unlike the others which are rendered *under* the current bullet, we show
it *in place* of the current bullet content. It kinda becomes the
defacto content for that object (even though technically it's just a set
of related objects).
- Convert relation list item "pinned" property to be an enum b/w
pinned/noteContent/all
- Update updateRelationList operation to support all types of lists
(pinned, note, all). Also add an add/move from relation list transaction
which generates those operations. (This can probably replace
pinRelation/unpinRelation but I haven't done this)
- When noteContentRelationList is empty, we render the node.content
prop, otherwise we render the noteContentRelationList bullets in it's
place.
- Remove bundle and zone. This has been dead code for a while. A lot of
the touched files come from this.
- Merge changes
  - Backspacing from note into note merges them
  - Backspacing from note into bullet adds the bullet to the note
  - General refactor so I could keep my head on straight
- Split changes
  - in outline view, shift+enter creates a note
  - in note view, enter by itself creates a note
  - cmd+enter splits a note

- When an object is has note content, use as search text the
concatenation of the text of all the related objects flagged as note
content.

Completes ENT-4108, ENT-3529

## How to test

- Watch the linked video and try out the new behaviours