Avoid infinite recursion in multiline note rendering (#590)
Multiline notes have a different code path for hydration which goes
around the isExpanded prop. This could lead to infinite recursion. For
now, we just don't hydrate if a given object is already in the path.
Later we'll figure out a ui for this.

Completes ENT-4281

## How to test

Create a multiline note and then reference itself
![CleanShot 2024-10-29 at 11 29
36@2x](https://github.com/user-attachments/assets/5a4b7ae8-b89e-4a32-9149-c4f17137cce1)
After selecting the option from the dropdown, the app will crash