Taylor/hardcode user (#592)
Allow devs to run the app as any user without login.

Add a `NEXT_PUBLIC_HARDCODED_USER_ID` env variable. Both the front and
backend have this.
 
If auth is disabled and the hardcoded user is set, then both the front
end and use that hardcoded user. The front end uses that id for
`fetchGetOrCreateUser` and sends unauthenticated requests. The backend
skips auth checks and assigns the hardcoded user as the requester. If
auth is enabled, then we go through the same flow as normal.

## How to test

- Pull this PR
- In your .env
- set `NEXT_PUBLIC_HARDCODED_USER_ID` to the mew user id you use in prod
  - set the postgres url to production
  - set `NEXT_PUBLIC_IS_AUTH_ENABLED` to false
  - set `NEXT_PUBLIC_PERSISTENCE_ENABLED` to true
- Run the app `yarn dev`
- It should load with the specified users data without prompting for
login