Fix Pusher connection logic (#605)
## Summary

We've been seeing way more concurrent Pusher connections than we should
be, and I was able to track it down to us re-instantiating the client
whenever we called `startSync()` in `UpdateManager`, which is in turn
called every time we do `resetAndLoad` — something that happens a fair
bit (especially in longer lived sessions).

Basically if a single user encountered a dozen sync errors that
triggered `resetAndLoad` over the course of a long session, they ended
up being a dozen concurrent connections.

Solution is to hoist out Pusher client instantiation into static space,
and so `unsubscribe` calls instead of `disconnect` so that we use a
single Pusher connection for the life of the client.

## How to test

Open dev tools, watch the network tab, do
`window.mew.graphStore.updateManager.fetchLatestDataSnapshot();` and
watch number of websocket connections being created. In `main` right now
you'll see a new one each time you make the call; in this branch you'll
only ever see one.