Add back copy of lexical typeahead plugin with modified positioning (#602)
When the user creates a mention inside a blue-outlined editor, it gets
positioned funny, jutting out to the left when you'd expect it to go to
the right (see screenshot on ticket). In the blue-outline state, the
editor fits the width of the content so that 1) we can put a blue
outline around it and 2) so we can put a place _outside_ the editor
where the user can put their focus. This is where the problem comes in:
the lexical typeahead plugin positions the dropdown, and it positions it
_relative to the editor bounding box_. It assumes the edge of the editor
is roughly the edge of the screen, so positions the dropdown to the left
to keep it within view. This results in the weird position in our case
cause the editor's right edge is near the middle of the screen.

Things I tried which didn't work:
1. Use css to reposition dropdown. The lexical typeahead plugin
repositions the dropdown in javascript. Fighting it was a losing battle.
2. Making editor full width and then using css which places the blue box
only around the editor text (instead of around the editor, which is now
much wider than the content). In this approach, you lose the input
suffix element which has special behaviour built into it e.g. pressing
backspace deletes the current node and replaces with a copy. With the
input element, we can positions the cursor *outside* the blue outline,
which is desired. I couldn't find a clever way to do that while keeping
the bullet one continuous editor.

I eventually decided to bring back a copy of lexical typeahead plugin
and slight modify how it does positioning (a one line change).

https://github.com/IdeaFlowCo/mew/blob/76e8e045af82a4e298a0c8496bce6e975d8df5c2/src/app/editor/plugins/dropdown/LexicalMenu.ts#L426

This feels silly, cause we already did this and then removed it in #477.
But that time we had deeply modified the code and coupled it with our
application state. I still think most of the refactoring in #477 was the
right call. The thing that I missed was that one of those changes was
necessary i.e. positioning the mention dropdown.

What I really _would_ like to do is remove the use of the lexical
typeahead plugin and do the positioning ourselves. We already need to do
some [funky
things](https://github.com/IdeaFlowCo/mew/blob/9037f944ebe269bf535b0d73cde7d3d55f355f5e/src/app/editor/plugins/dropdown/DropdownPlugin.tsx#L71-L90)
to make our dropdown plugin fit the typeahead interface. I expect we'd
be able to make the whole dropdown plugin code much simpler if we did
this.

That seems like a _lot_ of work though, especially since I think the
blue-outline state/design is a bad idea in the first place and [should
be
changed](https://linear.app/ideaflow/issue/ENT-3454/spec-blue-outline).
Why put in a bunch of work to address constraints that will be removed
later?

So for now I'm just copy pasting a bunch of code and doing the one line
change. The only alternative option is ignoring the problem (a real and
good option) but Jacob keeps complaining about this (see links on
ticket) so I'm going with this approach.

Completes ENT-3999