Expansion State Preservation when splitting Multiline Notes (#613)
## Summary

This fix does an in-order tree traversal to find all the expansions of
note, and caches them to apply to the new split of the multiline note.

The main issue I encountered was figuring out how to manipulate the path
of a given node so that I could use `setExpansionState` on the new path
of the multiline note. This was as simple as getting the suffix of the
path after `noteContent/` piece.

I'm not sure if the `noteContent` part of the path is something we plan
to keep though, so I think it's better to move the path manipulation
part to a separate class that handles all things path-related. I think
this is outside the scope of this PR though, perhaps something I'll
create a linear issue for?

What're your opinions on using recursion in TS? I used it because it was
easy but i could very easily change it to a loop instead, just let me
know.

I'm a bit worried that the recursion could lead to infinite recursion in
the event of a cycle, I'll be able to test this out because the
multiline note cycles are my next ticket. For now I have an infinite
recursion guard that stops the traversal whenever it re-encounters a
node - this has potentially weird edge cases like if the same id exists
elsewhere in the ml-note and so it doesn't get expanded. Better than a
crash though.

Completes ENT-4271

## How to test

Create some multi-line note with multiple levels of expansions like
this:

![screenshot](https://github.com/user-attachments/assets/fdbf1d08-3d2a-4bb1-9911-519859f398e1)

Try to split it in the middle. Expansions should be preserved.

## Checklist

- [X] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [X] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).