Multiline Note Backrelations UI Fix (#618)
## Summary

Completes ENT-4282

The problem was that we were checking for `!isBackRelation` in an `and`
statement in the `isUnlabelledChild` function. Not sure why exactly this
was, but as a result it would designate the circular ref as *not* an
unlabelled child. I checked and `isUnlabelledChild` is only used for
this one purpose in the entire codebase.

@taylormitchell is there some reason we wanted this in the past? Don't
want to ship a regression.

## How to test

Create a multiline note with:
- a
- b
- [semicolon] a b
and see that there is a circular reference that pops up, without the
"child" label.

## Checklist

- [X] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [ X I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).