Link Undoing (#620)
## Summary

The problem has to do with the way links are recognized. Looks like on
every render the link plugin attempts to find new links. When a new link
is recognized in the tree, an `updateNode` operation is done so that the
text is changed from mere text into a link.

However, when undoing, you undo the text->link operation (which is part
of a different transaction on the undo stack), so the link is back to
being normal text. When that happens, the render loop once again changes
the text to a link, hence the seemingly nothing behaviour.

Tried a few things that didn't work, and settled on adding a condition
in the `queueManager` that pushes link updates into the previous
transaction. That way the undo operation undoes everything we
intuitively want it to by undoing the most recent transaction. This
handles ugly cases like multiple lines of link and non-link content
pretty well.

I don't like that a node is updated when a link is recognized. I think
this kind of implicit behaviour should only be a display thing rather
than a change to the underlying structure. I think we should drill down
the exact use case of the link recognition feature (because I do see why
it's useful to change the structure with certain links) and make UI that
is more imperative than implicit.

Also added a new unit test.

Completes ENT-4228

## How to test

Copy paste the following into the editor and undo:
```
https://www.google.com/
A
https://youtube.com
B
C
D

```

## Checklist

- [X] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [X] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).