Split by dashes feature (#621)
## Summary

Now you can split notes using dashes. This only works on nodes that are
at the root rel. to notecontent.

Big toughie with this was figuring out how to deal with potential
children of the node being split on. That is, if we have node structure
like this
```
a bullet
---<--and then press enter here)
- 1
- 2
- 3
another bullet
```

In this case I opted to simply attach the children to the root of the
note content in the next note. Makes sense intuitively to me because
you're still splitting at the given line. Only thing I can think of that
would be better is to create an empty node to parent these orphans.

Edit: I ended up making an empty parent because it's easier to preserve
expansion states this way. Also is more elegant

Didn't write test cases because we may decide on different behaviour.

Completes ENT-3527 

## How to test

Create a multiline note and split it with `---`

## Checklist

- [X] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [X] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).