Implements note conversion on `Shift+Enter` for the TreeSuffix and Prefix's (#633)
## Summary

So it works, but there's some weird note logic happening. When you
convert the global (or reference) node to a note, with the way current
note content is converted and designated, the other global node is also
converted to a note. I think it would be more intuitive for it to become
a global node within the new note. I tried to do this quickly but it's
not easy, so I figured I'd wait for feedback before continuing.

Note: I refactored the existing `handleConvertToNote` so that the
majority of it happens within a function of `Tree`.

Edit: @taylormitchell and I chatted about this, concluded it's ok to
leave as-is for now.

Completes ENT-4301

## How to test

Create a node, and another node by referencing it. Convert one of them
to a MLNote by pressing `Shift-Enter` in the prefix or suffix of the
node.

## Checklist

- [X] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [X] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).