Parse with AI feature (#661)
## Summary

There's issues in here but this implementation basically works.

Biggest annoyance at the moment is just that there's no loading
indicator to tell you something is happening while the data is being
fetched from Open AI. Figured it's better to do that in a quick
follow-up PR just to make sure feature is live.

An inevitable next step here will be trying to match extracted entities'
text to autolink things into the graph, but that's kind of complicated
and something we should discuss. It's hard to decide the right line of
how aggressive to be with that.

We should also consider changing the prompt (which is copied over from
Thoughtstream) to be more Mew-y — e.g. specifying that relation targets
each have their own line etc. Without autolinking in place it didn't
seem as urgent as just getting this up.

~~Extra note: I wired this through our backend in order to protect the
Open AI API key, but didn't actually make it an authenticated endpoint
as it doesn't need user ID and doesn't interact directly with user data.
That probably shouldn't be that way but getting `authedFetch` into this
specific spot was more trouble than it was worth at this exact moment.~~
(Addressed this with Taylor's suggestion)

## How to test

Make a new node, paste some parseable text into it, try the new "Parse
with AI" button in the related object menu.

Sample text:
```
Theodore Roosevelt Jr.[b] (October 27, 1858 – January 6, 1919), often referred to as Teddy or T. R., was the 26th president of the United States, serving from 1901 to 1909. He previously was involved in New York politics, including serving as the state's 33rd governor for two years. He was the vice president under President William McKinley for six months in 1901, assuming the presidency after McKinley's assassination. As president, Roosevelt emerged as a leader of the Republican Party and became a driving force for anti-trust and Progressive policies. 
```

## Checklist

- [X] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [X] I tested my contribution.
- [X] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).