Loading indicator for parse with AI (#662)
Just a quick + dirty implementation here so that we have something.
Can/should be cleaned up later.

Note: I'm also including a quick fix for `StoresProvider` here to let us
use MOCK_MEW_USER when auth is disabled in `preview` environments, not
just `development`. This wasn't an issue before but we added the new
type possibility in #659 which ended up breaking the check we had in
place. (Previously, everything was either "development" or "production")