Add rag-based query system (#663)
This is a port of [Calvin's RAG based query
system](https://github.com/Calvin-Xu/mew-query/tree/main) written in
python to typscript. I mostly just copy-pasted, although it took a fair
amount of work to actually get it running. I simplified some things,
probably too much for Jacob's like, but it's working (and still too
overcomplicated imo).

Changes:
- Add a script which takes what db your .env is pointing at, vectorizes
it, and uploads it to
[pinecone](https://app.pinecone.io/organizations/-OCEnUiDq9I_39Pz4tEY/projects/72fc6e4a-6ff5-4d3f-8bbb-0db8e0dc7df1/indexes).
We use the database instance + name as a unique identifier for each set
of vectors (called indexes in pinecone).
- During serialization of the node, we don't just include the nodes
content, but also it's neighbours. Calvins implementation would do
multiple hops but this one only does one.
- There's an api end point which does the complicated stuff. It takes a
query and gets the K most similar nodes. Then chatgpt decides whether
it's ready to answer, or whether it wants to do another search. It can
do up to 5 search before it's forced to respond.
- On the client, there's a "query" page. It hits the query api, parses
the output, and renders it with ids included in the response linkified.

## How to test

- Pull repo and pull env variables
- Vectorize and upload whatever db is specified in your .env with `yarn
pinecone:upload`
- Run the app `yarn dev`
- In the sidebar, click "Query"
- Ask something about the current graph