Add sync status indicator (#680)
## Summary

Add a sync status indicator in top right to show if we have unsync'd
updates or not. Looks like this:

![image](https://github.com/user-attachments/assets/9f1795ff-f5c8-415b-ae4c-ee4507c9693a)

We can make this richer/more complicated later on but it's a starting
point. Future considerations could be things like watching for offline
errors in the `syncLoop` fetch within `UpdateManager` and reacting
accordingly.

Completes ENT-4194

## How to test

Make changes to data, notice indicator change status