Separate pusher channels with new env variable (#701)
## Summary

As discussed elsewhere, we currently have collisions in sync updates
being broadcast for public graphs that aren't stored in the same
database. This was mitigated by having separate Pusher configurations
for Development and Production, but is still an issue when developing
locally as we use different database instances. If Samuel is testing on
his database instance and makes an update to the public graph, my
browser pointing at my own test database will still get the update
broadcast to it on the global graph channel, which then causes an error
to be thrown when the object being updated/deleted/etc is inevitably
missing.

The solution in this PR is to add a new `.env` variable,
`NEXT_PUBLIC_PUSHER_CHANNEL_PREFIX`, that will be different in each
environment and developers can further change when they point to a
non-shared database. This new value is a prefix for each Pusher channel
ID (both the user channels and the global graph channel), giving us a
very easy lever to separate out Pusher communication between
environments.

Completes ENT-4406.

Something considered but not implemented here: trying to infer this
prefix automatically from the database specified in the Postgres
connection string. I decided that was ultimately more trouble than it
was worth because we need this prefix value available in both the client
and the backend and it'd be a big credential leak to make the connection
string public. That would mean we'd need to add something like a
`/config` endpoint to our API where the backend could serve this one
value computed from the connection string, and that just seemed like
overkill.

Might be worth doing in the future, though.

## How to test

Load locally (so that persistence is enabled) with dev tools open to
network tab. Look for the Pusher websocket connection. Make changes; see
that the channel ID has your specified prefix:


![image](https://github.com/user-attachments/assets/29d25085-8328-4d28-80db-e783a9c7a4d5)

## Checklist

- [X] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).