"My Hashtags" feature (#697)
## Summary

* Add in My Hashtags node that's created + pinned as part of
`ensureDefaultObjects` initialization
* Add new properties to `BaseGraphObject`, `isDeleteRestricted` and
`isEditRestricted` to centralize some of our delete/edit restriction
logic that we already had in place for Global Root and User Roots
* Make the new "My Hashtags" node restricted for both deletions and
updates
* Tweak @ mention node-creation logic to add newly created nodes under
"My Hashtags" when they start with `#`

Note: this does _not_ have a migration to add in a My Hashtags node to
every existing user. We could add one but it's not really necessary
since the `ensureDefaultObjects` initialization will do that
automatically whenever a user first logs in. It seems better to rely on
that then force stuff into the database, but I can add it in if
necessary.

Completes ENT-4242

## How to test

Login, create a new `#` prefixed node through the @ mention dropdown
(either in regular editor or command bar)

Make sure edit/deletion stuff looks reasonable