Optimistic syncs (#753)
## Summary

This was @nt7nguye 's idea - he mentioned we could just retry the sync a
few times whenever we get a 404.

Simple enough to implement so I tried it out. UX seems to be better.

## How to test

Play around with the app. Try to get it to fail a sync.

## Checklist

- [X] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [X] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).