fix crash when not-logged in (#763)
The app crashes if you log out right now. A change was made to the
command bar where it does a get on GraphStore.userRoot but when you're
not logged out, there is no user, so the getter throws. I think we're
gonna keep shooting ourselves in the foot if we have an interface that
suggests a user root always exists but then actually it can not exist in
a common case. For now, I'm just populating a dummy user node in the
case where we're not logged in.