Only convert to note at top level in note view
You create a new node inside the note-view. Initially, it's not a multiline note.

If the node has a labelled relation to the parent, we have a special case where pressing
enter will create a new multiline note and put the current related node inside it.

This is different from the usual behaviour for creating a multiline note, where the node
becomes a multiline note and the relation points to the whole multiline note. In the
special case above, the relation moves into the note too.

But we want this *only* for top-level nodes, but our code was doing it for nested nodes
too. This commit fixes that. Now it's only top-level.

Fixes bug at 25:10 of https://ideaflowteam.slack.com/archives/C07GW72L9RU/p1734950808648229