In the search index, index with and without non-alphanumeric prefixes (#764)
Keep the @[] wrapped around mentions in the search text. This ensures
"x" sorts higher than "@x" when you type "x" b/c matches closer to the
start are ranked higher.

We'd excluded those from the search text in #538 to address the fact
that our trie index only does prefix matching, so if you've indexed the
text "@x" then searches for "x" don't match, since your search text
isn't a prefix match. We move that handling inside the trie itself. If a
word starts with non-alphanumeric characters, we index it twice: once
as-is and once with those removed. So the node with text "@x" will be
indexed as "x" _and_ "@x". This was searches for "x" will match "@x".
This also gets us the behaviour requested in ENT-3864, where we want "x"
to match "#x" but we want want "#x" to match "#x". That last bit is why
we index twice: once with prefix included and once without.

Improve sorting so "x" appear higher than "@x" in the autocomplete
dropdown.
https://ideaflowteam.slack.com/archives/C07GW72L9RU/p1734950808648229

Completes ENT-3864

## How to test
- Create a node "x"
- Create another node "@x" and select "x"
- Create another node another node "@#x"
- Press `;` to open autocomplete and type x
- The autocomplete should contain all of the above, and it should be
sorted: x, #x, @x, @#x
- Make a new node and press `;` again, and type #x
- The autocomplete should be: #x, @#x