ENT-4567: Direct relation text is overlapping with pin (#771)
## Summary

Completes [ENT-4567: Direct relation text is overlapping with
pin](https://linear.app/ideaflow/issue/ENT-4567/direct-relation-text-is-overlapping-with-pin)

I spent a stupid amount of time on this but I actually learned a lot
about CSS modules here. I realized I've only used CSS-in-JS library like
Mantine so I've never used the fancy CSS operators like `:has`, `:not`,
`~`.

Getting to this CSS was kinda straighforward 

```
.RelatedObjectContent:hover:not(:has(.RelationCounter:hover)) .PinToggle {
```

but it wasn't working because the RelationCounter component is importing
from a different CSS module, which generates a different CSS class name
with unique id.

So I removed the `RelationCounter.module.css` as well, all the CSS
stylings are the same.

### Edit: While writing the PR desc, I realized I actually can just bump
the z-index of the tooltip and that works so much cleaner. Oh well.

## How to test



https://github.com/user-attachments/assets/48f6b584-11fd-4bbd-98c2-d250d11310fa



## Checklist

- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [ ] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).