Remove relation type with hotkey combo (#815)
## Summary

Mod-shift-backspace to remove a relationtype from an empty node.

Jacob wanted Mod-backspace. Doing it with Mod-backspace is really weird,
the browser interferes with the event structure. Here's what's
happening:

If we have it set to Mod-backspace, then when we are in a node like
this:
```
reltype - abc[cursor here]
```
and we press backspace, first the baked-in ctrl-backspace event
registers and clears abc:
```
reltype - [cursor here]
```
However, since this doesn't stop propagation, the ctrl-backspace hotkey
also triggers and since we're in an empty node with a relationtype, it
deletes the relationtype:
```
- [cursor here]
```

Any ideas?

## How to test

try pressing ctrl-shift-backspace in an empty node with a relationtype
on the relation with parent.

## Checklist

- [X] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [X] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).